# swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "REST API for Swagger Documentation",
        "description": "Local development server",
        "version": "1.0.0"
    },
    "components": {
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "name": "authorization",
                "in": "header",
                "description": "Authentication token"
            }
        }
    },
    "servers": [
        {
            "url": "https://utilities.divami.com/be/"
        }
    ],
    "paths": {
...
        "/connect": {
            "post": {
                "tags": [
                    "establishing the dn connections"
                ],
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "summary": "To connect to the db",
                "description": "establishing the connection with the db's",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "source": {
                                        "properties": {
                                            "type": {
                                                "type": "string",
                                                "example": ""
                                            },
                                            "host": {
                                                "type": "string",
                                                "example": ""
                                            },
                                            "databaseName": {
                                                "type": "string",
                                                "example": ""
                                            },
                                            "userName": {
                                                "type": "string",
                                                "example": ""
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": ""
                                            },
                                            "port": {
                                                "type": "string",
                                                "example": ""
                                            }
                                        }
                                    },
                                    "target": {
                                        "properties": {
                                            "type": {
                                                "type": "string",
                                                "example": ""
                                            },
                                            "host": {
                                                "type": "string",
                                                "example": ""
                                            },
                                            "databaseName": {
                                                "type": "string",
                                                "example": ""
                                            },
                                            "userName": {
                                                "type": "string",
                                                "example": ""
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": ""
                                            },
                                            "port": {
                                                "type": "string",
                                                "example": ""
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "message": "Database connected successfully",
                        "sourceSchema": {},
                        "targetSchema": {},
                        "status": "true"
                    },
                    "400": {
                        "message": "Invalid credentials for source",
                        "status": "false",
                        "code": 400
                    }
                }
            }
        },
        "/comparedata": {
            "post": {
                "tags": [
                    "comparing the schema"
                ],
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "summary": "To comare the schema",
                "description": "comparig the schema queries",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "caseSensitive": {
                                        "type": "bool",
                                        "example": true
                                    },
                                    "sourceDBQuery": {
                                        "type": "string",
                                        "example": "select * from test;"
                                    },
                                    "sourceDBType": {
                                        "type": "string",
                                        "example": "postgres"
                                    },
                                    "swap": {
                                        "type": "string",
                                        "example": "true"
                                    },
                                    "targetDBQuery": {
                                        "type": "string",
                                        "example": "select * from test;"
                                    },
                                    "targetDBType": {
                                        "type": "string",
                                        "example": "postgres"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "message": null,
                        "mismatches": [
                            {
                                "Serial No": 10,
                                "Index": 5,
                                "userid": "1bA7A3dc874da3c",
                                "firstname": "Lori",
                                "lastname": "Todd",
                                "sex": "Male",
                                "email": "buchananmanuel@example.net",
                                "phone": "689-207-3558x7233",
                                "dob": "12/1/1938",
                                "title": "Veterinary surgeon"
                            }
                        ],
                        "misMatchFile": "csvMissedsourceCsvudaybhaskar.talari@divami.com8cc395c94847957f2cef16ab6ca3f7c995ca67c0231d38203db8ab087c0b11082024-05-14T041745018Z.json",
                        "numberOfRecords": 1,
                        "status": false,
                        "code": 200
                    }
                }
            }
        },
...
    }
}